battle field technical scanner/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
Pillow
timm
scipy
datasets>=2.14.0
transformers>=4.53.2
accelerate>=0.21.0
evaluate